Use response macros and return success response from controller

+++ app/Modules/Authentication/Services/AuthenticateService.php
@@ -2,89 +2,106 @@
 
 namespace App\Modules\Authentication\Services;
 
-use App\Exceptions\ApiResponseException;
-use App\Modules\Donor\Services\GetDonorService;
-use App\Modules\Donor\Services\UpdateDonorService;
-use App\Modules\User\Services\CreateUserService;
-use App\Modules\User\Services\GetUserService;
-use App\Modules\User\Services\UpdateUserService;
+use App\Modules\DonationCore\Donor\Services\DonorService;
+use App\Modules\User\Exceptions\UserNotFoundException;
+use App\Modules\User\Resources\UserResource;
+use App\Modules\User\Services\UserService;
 use Carbon\Carbon;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\ValidationException;
 use MikeMcLin\WpPassword\Facades\WpPassword;
 
 class AuthenticateService
 {
     public function __construct(
-        private readonly GetUserService     $getUserService,
-        private readonly CreateUserService  $createUserService,
-        private readonly UpdateUserService  $updateUserService,
-        private readonly GetDonorService    $getDonorService,
-        private readonly UpdateDonorService $updateDonorService
+        private readonly UserService  $userService,
+        private readonly DonorService $donorService,
     )
     {
     }
 
+
+    /**
+     * @param array $credentials
+     * @return array
+     * @throws ValidationException
+     */
     public function login(array $credentials): array
     {
-        $user = $this->getUserService->getUserByEmailOrUsername($credentials['email']);
+        $user = $this->userService->getUserByEmailOrUsername($credentials['email']);
+
+        if (!$user) {
+            throw ValidationException::withMessages([
+                'email' => 'Invalid credentials',
+            ]);
+        }
 
         // Validate credentials and active status
         if ($user->active && WpPassword::check($credentials['password'], $user->password)) {
-            $this->updateUserService->updateUserPassword($user, $credentials['password']);
+            $user->password = Hash::make($credentials['password']);
+            $user->save();
         } else if (!Auth::attempt(['email' => $credentials['email'], 'password' => $credentials['password'], 'active' => '1'])) {
             throw ValidationException::withMessages([
                 'email' => 'Invalid credentials',
             ]);
         }
 
-        try {
-            return $this->generateTokenResponse($user);
-        } catch (\Exception $e) {
-            throw new ApiResponseException('An error occurred during the login process: ' . $e->getMessage());
-        }
+        return $this->generateTokenResponse($user);
     }
 
+
+    /**
+     * @param array $credentials
+     * @return array
+     */
     public function register(array $credentials): array
     {
-        $user = $this->createUserService->createUser($credentials);
+        $user = $this->userService->createUser($credentials);
 
         // Assign user to donor if existing
-        $donor = $this->getDonorService->getDonorByEmail($credentials['email']);
+        $donor = $this->donorService->getDonorByEmail($credentials['email']);
 
         if ($donor) {
-            $this->updateDonorService->assignDonorToUser($donor, $user->id);
+            $donor->user_id = $user->id;
+            $donor->save();
         }
 
-        try {
-            // Send email verification
-            event(new Registered($user));
+        // Send email verification
+        event(new Registered($user));
 
-            return $this->generateTokenResponse($user);
-        } catch (\Exception $e) {
-            throw new ApiResponseException('An error occurred during the register process: ' . $e->getMessage());
-        }
+        return $this->generateTokenResponse($user);
     }
 
+
+    /**
+     * @return bool
+     * @throws UserNotFoundException
+     */
     public function logout(): bool
     {
-        $user = $this->getUserService->getAuthUser();
+        $user = $this->userService->getAuthUser();
 
-        try {
-            return $user->currentAccessToken()->delete();
-        } catch (\Exception $e) {
-            throw new ApiResponseException('An error occurred during the logout process: ' . $e->getMessage());
+        if (!$user) {
+            throw new UserNotFoundException();
         }
+
+        return $user->currentAccessToken()->delete();
     }
 
+
+    /**
+     * @param $user
+     * @return array
+     */
     private function generateTokenResponse($user): array
     {
         $token = $user->createToken('apiToken')->plainTextToken;
         $tokenExpiresAt = Carbon::now()->addMinutes(config('sanctum.expiration'));
 
         return [
-            'user' => $user,
+            'user' => new UserResource($user),
             'accessToken' => $token,
             'tokenExpiresAt' => $tokenExpiresAt->getTimestampMs(),
         ];
